refactor(front): move workflows core table under object-core
Sits next to object-record and object-metadata, with room for agents and
sequences tables alongside workflows.

Reverts the useSortedArray date sorting: updatedAt is an ISO string, which
already sorts chronologically as a string.
