+++ packages/twenty-front/src/modules/object-core/workflows/constants/WorkflowCoreStatusTag.ts
@@ -1,7 +1,7 @@
 import { type MessageDescriptor } from '@lingui/core';
 import { msg } from '@lingui/core/macro';
 
-import { type CoreWorkflowStatus } from '@/workflow-core/types/CoreWorkflow';
+import { type CoreWorkflowStatus } from '@/object-core/workflows/types/CoreWorkflow';
 
 export const WORKFLOW_CORE_STATUS_TAG: Record<
   CoreWorkflowStatus,
