+++ packages/twenty-front/src/modules/object-core/workflows/graphql/queries/getCoreWorkflows.ts
