+++ packages/twenty-front/src/modules/object-core/workflows/hooks/useCoreWorkflows.ts
@@ -1,8 +1,8 @@
 import { useQuery } from '@apollo/client/react';
 
 import { useApolloCoreClient } from '@/object-metadata/hooks/useApolloCoreClient';
-import { GET_CORE_WORKFLOWS } from '@/workflow-core/graphql/queries/getCoreWorkflows';
-import { type CoreWorkflow } from '@/workflow-core/types/CoreWorkflow';
+import { GET_CORE_WORKFLOWS } from '@/object-core/workflows/graphql/queries/getCoreWorkflows';
+import { type CoreWorkflow } from '@/object-core/workflows/types/CoreWorkflow';
 
 export const useCoreWorkflows = ({ skip }: { skip: boolean }) => {
   const apolloCoreClient = useApolloCoreClient();
