+++ packages/twenty-front/src/modules/object-core/workflows/types/CoreWorkflow.ts
