+++ packages/twenty-front/src/modules/ui/layout/table/types/TableFieldMetadata.ts
@@ -4,7 +4,7 @@ import { type IconComponent } from 'twenty-ui/icon';
 export type TableFieldMetadata<ItemType> = {
   fieldLabel: MessageDescriptor;
   fieldName: keyof ItemType;
-  fieldType: 'string' | 'number' | 'date';
+  fieldType: 'string' | 'number';
   align: 'left' | 'right';
   FieldIcon?: IconComponent;
 };
